backend/requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv

faiss-cpu

google-generativeai
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters

openpyxl
pandas
pdfplumber
python-docx
python-pptx
